Organize MenuBar components
